+++ .gitignore
@@ -12,6 +12,7 @@
 # Output of the go coverage tool, specifically when used with LiteIDE
 *.out
 
+.eslintcache
 data/
 dist/
 node_modules/
